@@ -12,8 +12,8 @@ license = "MIT"
 "Release Notes" = "https://github.com/langchain-ai/langchain/releases?q=tag%3A%22__package_name_short__%3D%3D0%22&expanded=true"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
-langchain-core = "^0.2.0"
+python = ">=3.9,<4.0"
+langchain-core = "^0.3.0.dev"
 
 [tool.poetry.group.test]
 optional = true
