@@ -6,8 +6,8 @@ authors = []
 readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">=3.8.1,<4.0"
-langchain-core = ">=0.1.5,<0.3"
+python = ">=3.9,<4.0"
+langchain-core = "^0.3.0.dev"
 langchain-openai = ">=0.0.1"
 
 
