@@ -1,5 +1,5 @@
 [build-system]
-requires = [ "poetry-core>=1.0.0",]
+requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
@@ -12,7 +12,7 @@ repository = "https://github.com/langchain-ai/langchain"
 license = "MIT"
 
 [tool.ruff]
-select = [ "E", "F", "I",]
+select = ["E", "F", "I"]
 
 [tool.mypy]
 disallow_untyped_defs = true
@@ -23,20 +23,24 @@ disallow_untyped_defs = true
 
 [tool.poetry.dependencies]
 python = ">=3.9,<4.0"
-langchain-core = ">=0.1.52,<0.3"
+langchain-core = "^0.3.0.dev"
 qdrant-client = "^1.10.1"
 pydantic = "^2.7.4"
-fastembed = { version = "^0.3.3", python = ">=3.8.1,<3.13", optional = true}
+fastembed = { version = "^0.3.3", python = ">=3.9,<3.13", optional = true }
 
 [tool.poetry.extras]
-fastembed = [ "fastembed"]
+fastembed = ["fastembed"]
 
 [tool.coverage.run]
-omit = [ "tests/*",]
+omit = ["tests/*"]
 
 [tool.pytest.ini_options]
 addopts = "--snapshot-warn-unused --strict-markers --strict-config --durations=5"
-markers = [ "requires: mark tests as requiring a specific library", "asyncio: mark tests as requiring asyncio", "compile: mark placeholder test used to compile integration tests without running them",]
+markers = [
+    "requires: mark tests as requiring a specific library",
+    "asyncio: mark tests as requiring asyncio",
+    "compile: mark placeholder test used to compile integration tests without running them",
+]
 asyncio_mode = "auto"
 
 [tool.poetry.group.test]
